Fix: Empty click gui at first launch
